Optimize set(), unset(), & get() - y-vars to use only 8-bits.

+++ src/life.h
@@ -3,9 +3,11 @@
 #include <stdint.h>
 
 
-void    set(int x, int y);
-void  unset(int x, int y);
-uint8_t get(int x, int y);
+/* optimize y-var in calls by using 8-bits since y<240 */
+void    set(int x, uint8_t y);
+void  unset(int x, uint8_t y);
+uint8_t get(int x, uint8_t y);
+
 void setup(void);
 void next(void);
 void main(void);
